refactor(pipeline): minor improvements to filenames and structure
