+++ src/core/print/print.ts
@@ -1,4 +1,4 @@
-import type { IPipelineStep, StepProcessingFunction } from '../steps/step'
+import type { IPipelineStep, StepProcessingFunction } from '../steps/pipeline-step'
 import type { TailOfArray } from '../../util/arrays'
 import { guard } from '../../util/assert'
