+++ src/core/steps/all/core/00-parse.ts
@@ -1,7 +1,7 @@
 import { internalPrinter, StepOutputFormat } from '../../../print/print'
 import { parseToQuads } from '../../../print/parse-printer'
-import type { IPipelineStep } from '../../step'
-import { PipelineStepStage } from '../../step'
+import type { IPipelineStep } from '../../pipeline-step'
+import { PipelineStepStage } from '../../pipeline-step'
 import type { RParseRequest, RShell } from '../../../../r-bridge'
 import type { DeepReadonly } from 'ts-essentials'
 import type { RShellExecutor } from '../../../../r-bridge/shell-executor'
