+++ src/core/steps/all/core/10-normalize.ts
@@ -9,8 +9,8 @@ import {
 	printNormalizedAstToMermaid,
 	printNormalizedAstToMermaidUrl
 } from '../../../print/normalize-printer'
-import type { IPipelineStep } from '../../step'
-import { PipelineStepStage } from '../../step'
+import type { IPipelineStep } from '../../pipeline-step'
+import { PipelineStepStage } from '../../pipeline-step'
 import type { DeepReadonly } from 'ts-essentials'
 import type { ParseRequiredInput } from './00-parse'
 import { normalize } from '../../../../r-bridge/lang-4.x/ast/parser/json/parser'
