+++ src/core/steps/all/static-slicing/00-slice.ts
@@ -1,6 +1,6 @@
 import { internalPrinter, StepOutputFormat } from '../../../print/print'
-import type { IPipelineStep } from '../../step'
-import { PipelineStepStage } from '../../step'
+import type { IPipelineStep } from '../../pipeline-step'
+import { PipelineStepStage } from '../../pipeline-step'
 import type { SlicingCriteria } from '../../../../slicing'
 import { staticSlicing } from '../../../../slicing'
 import type { DeepReadonly } from 'ts-essentials'
