+++ src/core/steps/pipeline/default-pipelines.ts
