+++ src/core/steps/pipeline/index.ts
@@ -1,3 +1,3 @@
 export * from './pipeline'
 export * from './invalid-pipeline-error'
-export * from './default'
+export * from './default-pipelines'
