+++ src/core/steps/pipeline/pipeline.ts
@@ -1,5 +1,5 @@
-import type { IPipelineStep, PipelineStepName, PipelineStepStage } from '../step'
-import { verifyAndBuildPipeline } from './create'
+import type { IPipelineStep, PipelineStepName, PipelineStepStage } from '../pipeline-step'
+import { verifyAndBuildPipeline } from './create-pipeline'
 import type { DeepReadonly, UnionToIntersection } from 'ts-essentials'
 
 /**
